LF-13203 fix(imx93): init wakeupmix/nocmix always on by default
In current M33 power mode support demo, it does not claim the wakeupmix
and nocmix on/off, if the M33 enter suspend mode, these mix may be off wrongly.
So init wakeupmix & nocmix to always on for each domain. when a CPU domain
want to control wakeupmix or nocmix, it should set the lpm whitelist & lpm
setting to expected state.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>